smartsys/requirements.txt
jsonschema
paho-mqtt
psycopg2
python-telegram-bot
PyYAML
RPi.bme280
RPi.GPIO